Introduced actualTime and scheduledTime instead of having different arrival and departure times
Minor change:
Since a plane performs only one action at a time, thought it would be better of we reduced those two categories into just one i.e. time.

Another suggestion: Should we introduce isTakenOff as well as it is counterpart to isLanded

Concern: Had this concern about how someone would know what the actual time is when creating the object so was wondering if we could initialize actual value in the constructor on our own?